requirements.txt
langgraph
langchain
langchain-community
langchain-chroma
sentence-transformers
chromadb
groq
mailslurp-client
streamlit
python-dotenv
pydantic
